Revert "Add separate pair screen"
This reverts commit 53cc93d2def29212010f627647d8fc01f68ab2fe.
