+++ app/src/main/AndroidManifest.xml
@@ -35,9 +35,6 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".views.PairActivity"
-            android:label="@string/pair" />
-        <activity
             android:name=".views.HelpActivity"
             android:label="@string/more" />
         <activity
