+++ app/src/main/java/com/draco/ladb/views/MainActivity.kt
@@ -1,6 +1,8 @@
 package com.draco.ladb.views
 
+import android.content.ComponentName
 import android.content.Intent
+import android.net.Uri
 import android.os.Bundle
 import android.view.KeyEvent
 import android.view.Menu
@@ -12,6 +14,7 @@ import android.view.inputmethod.InputMethodManager
 import android.widget.ScrollView
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.activity.viewModels
+import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.content.FileProvider
 import androidx.lifecycle.lifecycleScope
@@ -19,7 +22,9 @@ import com.draco.ladb.BuildConfig
 import com.draco.ladb.R
 import com.draco.ladb.databinding.ActivityMainBinding
 import com.draco.ladb.viewmodels.MainActivityViewModel
+import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import com.google.android.material.snackbar.Snackbar
+import com.google.android.material.textfield.TextInputEditText
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlin.system.exitProcess
@@ -28,6 +33,8 @@ class MainActivity : AppCompatActivity() {
     private val viewModel: MainActivityViewModel by viewModels()
     private lateinit var binding: ActivityMainBinding
 
+    private lateinit var pairDialog: MaterialAlertDialogBuilder
+
     private var lastCommand = ""
 
     private var bookmarkGetResult = registerForActivityResult(ActivityResultContracts.StartActivityForResult()) {
@@ -35,29 +42,15 @@ class MainActivity : AppCompatActivity() {
         binding.command.setText(text)
     }
 
-    private var pairGetResult = registerForActivityResult(ActivityResultContracts.StartActivityForResult()) { result ->
-        val port = result.data?.getStringExtra("port") ?: ""
-        val code = result.data?.getStringExtra("code") ?: ""
-
-        viewModel.adb.debug("Trying to pair...")
-        lifecycleScope.launch(Dispatchers.IO) {
-            val success = viewModel.adb.pair(port, code)
-
-            if (success) {
-                viewModel.setPairedBefore(true)
-                viewModel.startADBServer()
-            } else {
-                /* Failed; try again! */
-                viewModel.setPairedBefore(false)
-                viewModel.adb.debug("Failed to pair! Trying again...")
-                runOnUiThread { pairAndStart() }
-            }
-
-            viewModel.isPairing.postValue(false)
-        }
-    }
-
     private fun setupUI() {
+        pairDialog = MaterialAlertDialogBuilder(this)
+            .setTitle(R.string.pair_title)
+            .setCancelable(false)
+            .setView(R.layout.dialog_pair)
+            .setPositiveButton(R.string.pair, null)
+            .setNegativeButton(R.string.help, null)
+            .setNeutralButton(R.string.settings, null)
+
         binding.command.setOnKeyListener { _, keyCode, keyEvent ->
             if (keyCode == KeyEvent.KEYCODE_ENTER && keyEvent.action == KeyEvent.ACTION_DOWN) {
                 sendCommandToADB()
@@ -124,8 +117,16 @@ class MainActivity : AppCompatActivity() {
     private fun pairAndStart() {
         if (viewModel.needsToPair()) {
             viewModel.adb.debug("Requesting pairing information")
-            viewModel.isPairing.value = true
-            pairGetResult.launch(Intent(this, PairActivity::class.java))
+            askToPair { thisPairSuccess ->
+                if (thisPairSuccess) {
+                    viewModel.setPairedBefore(true)
+                    viewModel.startADBServer()
+                } else {
+                    /* Failed; try again! */
+                    viewModel.adb.debug("Failed to pair! Trying again...")
+                    runOnUiThread { pairAndStart() }
+                }
+            }
         } else {
             viewModel.startADBServer()
         }
@@ -142,6 +143,68 @@ class MainActivity : AppCompatActivity() {
             pairAndStart()
 
         viewModel.piracyCheck(this)
+    }
+
+    /**
+     * Ask the user to pair
+     */
+    private fun askToPair(callback: ((Boolean) -> (Unit))? = null) {
+        pairDialog
+            .create()
+            .apply {
+                setOnShowListener {
+                    getButton(AlertDialog.BUTTON_POSITIVE).setOnClickListener {
+                        val port = findViewById<TextInputEditText>(R.id.port)!!.text.toString()
+                        val code = findViewById<TextInputEditText>(R.id.code)!!.text.toString()
+                        dismiss()
+
+                        lifecycleScope.launch(Dispatchers.IO) {
+                            viewModel.adb.debug("Trying to pair...")
+                            val success = viewModel.adb.pair(port, code)
+                            callback?.invoke(success)
+                        }
+                    }
+
+                    getButton(AlertDialog.BUTTON_NEUTRAL).setOnClickListener {
+                        val intent = Intent(Intent.ACTION_MAIN).setComponent(
+                            ComponentName(
+                                "com.android.settings",
+                                "com.android.settings.Settings\$DevelopmentSettingsDashboardActivity"
+                            )
+                        )
+                            .addFlags(Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_LAUNCH_ADJACENT)
+
+                        try {
+                            startActivity(intent)
+                        } catch (e: Exception) {
+                            e.printStackTrace()
+                            Snackbar.make(
+                                binding.output,
+                                getString(R.string.snackbar_intent_failed),
+                                Snackbar.LENGTH_SHORT
+                            )
+                                .show()
+                        }
+                    }
+
+                    getButton(AlertDialog.BUTTON_NEGATIVE).setOnClickListener {
+                        val intent = Intent(Intent.ACTION_VIEW, Uri.parse(getString(R.string.tutorial_url)))
+                        try {
+                            startActivity(intent)
+                        } catch (e: Exception) {
+                            e.printStackTrace()
+                            Snackbar.make(
+                                binding.output,
+                                getString(R.string.snackbar_intent_failed),
+                                Snackbar.LENGTH_SHORT
+                            )
+                                .show()
+                        }
+
+                    }
+                }
+            }
+            .show()
     }
 
     override fun onOptionsItemSelected(item: MenuItem): Boolean {
